Split libmkcg into different files
Move each exported function into a separate file.
