+++ configure.ac
@@ -31,7 +31,7 @@ AC_INIT([PC/M Tools],
 AC_CANONICAL_BUILD
 AC_CANONICAL_HOST
 
-AC_CONFIG_SRCDIR([src/mkcg.c])
+AC_CONFIG_SRCDIR([src/mkcg_cli.c])
 AC_CONFIG_HEADERS([config.h])
 AC_CONFIG_MACRO_DIR([m4])
