+++ src/Makefile.am
@@ -21,7 +21,7 @@
 
 AM_CPPFLAGS = -I$(top_srcdir)/src
 
-check_PROGRAMS = 			\
+check_PROGRAMS =			\
 	mkcg.6416			\
 	mkcg.vid2k
 
@@ -29,13 +29,18 @@ check_PROGRAMS = 			\
 noinst_LTLIBRARIES = libmkcg.la
 bin_PROGRAMS = mkcg.6416 mkcg.vid2k
 
-mkcg_6416_SOURCES = mkcg.6416.c
+mkcg_6416_SOURCES = mkcg_cmd_6416.c
 mkcg_6416_LDADD = libmkcg.la
 
-mkcg_vid2k_SOURCES = mkcg.vid2k.c
+mkcg_vid2k_SOURCES = mkcg_cmd_vid2k.c
 mkcg_vid2k_LDADD = libmkcg.la
 
-libmkcg_la_SOURCES = mkcg.c
+libmkcg_la_SOURCES =			\
+	mkcg_isvalidch.c		\
+	mkcg_out_banner.c		\
+	mkcg_out_xxd.c			\
+	mkcg_out_xpm.c			\
+	mkcg_cli.c
 libmkcg_la_CFLAGS = $(AM_CFLAGS) @XPM_CFLAGS@
 libmkcg_la_LDFLAGS = $(AM_LDFLAGS) @XPM_LIBS@
